add mroe info to cmap
